requirements.txt
pandas
geopandas
jupyter
pyspark
rioxarray
apache-sedona[spark]
xarray[io]
matplotlib
exactextract
numpy<2.0
pyarrow
pydantic
exactextract
arlas_cli
